refactor: replace code to a new component cart from super-nav
