+++ src/components/pages/shop/cart/Cart.css
@@ -1,0 +1,7 @@
+.cart_wrap {
+    display: flex;
+    justify-content: space-between;
+    width: 150px;
+    cursor: pointer;
+    margin: 0 1rem;
+}
